Cleanup unsigned types on api (#9)
* cleanup the experimental flags that are useless

* cleanup the unsigned types on the api to make it simpler

+++ src/androidMain/kotlin/com/ditchoom/buffer/BufferFactory.kt
@@ -7,19 +7,19 @@ import android.os.SharedMemory
 import java.nio.ByteBuffer
 
 actual fun PlatformBuffer.Companion.allocate(
-    size: UInt,
+    size: Int,
     byteOrder: ByteOrder
 ): PlatformBuffer {
     val byteOrderNative = when (byteOrder) {
         ByteOrder.BIG_ENDIAN -> java.nio.ByteOrder.BIG_ENDIAN
         ByteOrder.LITTLE_ENDIAN -> java.nio.ByteOrder.LITTLE_ENDIAN
     }
-    return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O_MR1 && size > 0u) {
-        val sharedMemory = SharedMemory.create(null, size.toInt())
+    return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O_MR1 && size > 0) {
+        val sharedMemory = SharedMemory.create(null, size)
         val buffer = sharedMemory.mapReadWrite().order(byteOrderNative)
         ParcelableSharedMemoryBuffer(buffer, sharedMemory)
     } else {
-        JvmBuffer(ByteBuffer.allocateDirect(size.toInt()).order(byteOrderNative))
+        JvmBuffer(ByteBuffer.allocateDirect(size).order(byteOrderNative))
     }
 }
 
@@ -34,4 +34,4 @@ actual fun PlatformBuffer.Companion.wrap(array: ByteArray, byteOrder: ByteOrder)
 
 actual fun String.toBuffer(): PlatformBuffer = JvmBuffer(ByteBuffer.wrap(encodeToByteArray()))
 
-actual fun String.utf8Length(): UInt = encodeToByteArray().size.toUInt()
+actual fun String.utf8Length(): Int = encodeToByteArray().size
